Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-2
> 
> This code introduces the genuine possibility that ino_t numbers might change. Quite a lot of code in the tree assumes, at present, that will never happen (e.g. entry_refs will no longer work if the directory ino_t changes.)
> 
> Were the implications of this thought through very carefully at the time this code was originally done? Because, it seems to me to have quite the potential to break a running system in some very strange ways if applications depend on ino_ts like this (if they have open FDs, those should be fine as long as VFS is informed of the change, which it is.)
> 
> CCing axeld, who may remember something about this.

This is meant to be solved by a latter commit in this patch series which makes ino_t partially independant from the underlying disk offset: https://review.haiku-os.org/c/haiku/+/953/10

This allows the resizing code to set up an indirection for the inodes that are moved.

Note that this code path is only used when shrinking a partition, so if this part seems too dangerous, we could refactor (once again) this patch series to first merge the parts that allow to implement growing filesystems only. It may be a little simpler to review.

Also note that this specific commit does not call the code to move inodes, and as such, it should be safe to merge on its own. Potential problems happen in later commits which actually make use of this.

If the "moving inodes" part seems not ready enough, we could also refactor this patch series (again) to separate the code that only allows to grow a filesystem. When growing, there is no need to move any inodes, and that is probably the most useful use case for filesystem resizing.

Patch-set: 9
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>