Update patch set 3
Patch Set 3:

Indeed there could be security implications.

Some ideas for doing this in a non-intrusive way:

> I forgot to write how the security issue with accessing the clipboard is resolved in far2l. far2l allows you to read the clipboard using special escape sequence only if within the last 5 seconds user pressed Ctrl+C or right-clicked the mouse. If some script tries to read the clipboard on its own initiative, nothing will come of it.

(from the KDE bugreport linked)

Another possibility is maybe allowing it only for local apps, but there may not be a reliable way to detect that. And sharing clipboard with remote apps through ssh can be a cool feature too.

I'm a bit undecided. Is it worth exposing the option at all if we are not comfortable enabling it by default? Personally I don't think I have needed it in the past (Terminal own copy/paste features worked well enough).

Patch-set: 3