Update patch set 5
Patch Set 5:

Ah, ok. Following the BIOS preprogrammed setup is something I added myself as well, in the general version of setting up pipes (pipepreferences). This solved many issues in practice since reprogramming simply does not work reliably for FDI systems.

So I think that would be fine for LVDS as well. Only I cannot test this myself since I don't have the specific hardware for that. 
If you want to follow BIOS setup for LVDS specifically with this patch, then maybe it's a good plan to create a new patch for that based on the current state of the driver.
If following BIOS for LVDS would already fit in the 'general' case, it could even be moved there for LVDS as well I think?

Patch-set: 5
Reviewer: Gerrit User 1000057 <1000057@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>