Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > This is meant to be solved by a latter commit in this patch series which makes ino_t partially independant from the underlying disk offset: https://review.haiku-os.org/c/haiku/+/953/10
> 
> It looks like that just makes sure ino_t are not directly block numbers to catch any direct usages without conversions, but instead are convertible to block numbers via XORs? That does not make them any more independent...

The patch is unfinished, I think the idea of the XOR was to clearly differenciate between real inodes and userspace references to identify where the code needed changes. We don't even need to keep the XOR in the final version.

In the case of shrinking a filesystem, the old values for the remapped inodes would end up outside of the filesystem, so they are not hard to identify. The code can look something like this:

   if (ino_t < filesystem_end) {
       // use inode directly
   } else {
       // inode was moved, find the new position
       // bsearch in a fixed array of moved nodes should be possible, or use a map.
   }

In that case the overhead remains quite reasonable in the normal case (an extra check on an integer range, which you probably want to check anyway because trying to access data outside of a partition is a bad idea). It is a bit higher for ino_t values that got to userspace before the resizing was started.

Things would get more complicated for defragmentation, where it will be more difficult to identify moved inodes.

Patch-set: 9
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>