Update patch set 2
Patch Set 2:

(1 comment)

> rm -rf "objects.x86_64-cc11-release"
> rm: cannot remove 'objects.x86_64-cc11-release': Directory not empty

I have some directories it occurs on too. It may be due to FS corruption in that the directory is empty but libntfs-3g doesn't actually check that it is.

I have suspicions that the same problem will arise when trying to delete these directories on Linux. But I haven't yet proven that either way.

Patch-set: 2