Update patch set 2
Patch Set 2:

> Also I got a lot of ghost files that were not removed on shutdown and were visible on Windows.

Ghost files or directories? Because if they are directories, then the problem is the same as the "Directory not empty" you list above. Ghost files would be a separate issue.

Anyway, at least here, I can't delete some of those ghost files even in Linux under NTFS-3G. So it seems this is not a problem with our code anyway.

Patch-set: 2