Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> It looks like the same -boot change is needed for the non-cross repos:
> https://github.com/haikuports/haikuports/blob/master/sys-devel/gcc/gcc-11.2.0_2021_07_28.recipe
> https://github.com/haikuports/haikuports/blob/master/sys-devel/gcc/gcc-8.3.0_2019_05_24.recipe
> 

I took a look at the gcc-8.3.0 recipe in (non-cross) haikuports, it does not seem very difficuly to adapt the changes from haikuports.cross.

I was able to build it with boot libs already on x86_64, next: verify it on x86 secondary-arch as well. This latter one might take a bit longer as I'm building in a 32-bit Haiku VM in VirtualBox.

Patch-set: 2