Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> We now have a working FDT bus manager from the RISCV64 work. Perhaps it's time to get rid of the config_manager altogether?

I suppose so.
The interest of the config_manager was to allow switching between existing configurations in a preferences in BeOS, but it's not really implemented, and there's like no single driver left that uses it…
I think there's been some work for making part of FDTs optional or conditional, that might replace this function someday.

Patch-set: 2