Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I installed the Change-Id hook but it still refuge my change when I push?
> 
> Should the change id be autogenerated or should I use the Id above from Gerrit?

This change does not have a change id in the commit message at the moment, so Gerrit generated one when you first pushed it. You need to copy that change-id to your local commit now.

What the hook does is generate a change-id immediately locally whenever you create a commit. This way, you don't need to do this manual step.

> 
> Should I still push to master? Maybe that's what's wrong?

Changes should be pushed to refs/for/master to create or update change requests in Gerrit. Only developers with commit access can push directly to master to bypass the review process.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>