Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1: Code-Review-1
> 
> Indeed, average frequency between calls will probably be quite low due to cpuidle causing frequency reduction.

That's the point of this way of measuring frequency. For the max frequency of the CPU we already get the information in other ways. This is more about "how fast my CPU is currently running?"

The main usecase will be to plot it in ActivityMonitor, for example. In that specific case, this change works, as ActivityMonitor would be calling the function periodically. I think it would be the case for most applications using this function, it makes no sense to call it just once since it provides realtime information.

I think it is nice to let the application decide how often it wants to call the function, and this way adjust the integration period. Or even call the function as a profiling tool around a very specific piece of code (if you want to evaluate the effect of AVX512 on CPU frequency on Intel CPUs, for example, where they slow down the CPU because otherwise it would use too much power).

A similar approach was proposed for Linux:

"
MHz is the average frequency of the busy processor
over a measurement interval.  The interval is
defined to be the time between successive reads
of the frequency on that processor, whether from
/proc/cpuinfo or from sysfs cpufreq/scaling_cur_freq.
As with previous methods of calculating MHz,
idle time is excluded.

Note that frequent or concurrent reads of /proc/cpuinfo
or sysfs cpufreq/scaling_cur_freq will shorten the
measurement interval seen by each reader.  The code
mitigates that issue by caching results for 100ms.
"

(from https://patchwork.kernel.org/project/linux-pm/patch/2433e0a2ace56242bd67fd4ea13448125b1ed824.1459975343.git.len.brown@intel.com/)

Some research shows other appoaches were proposed such as doing this at every context switch (but it's costly). I think this one is reasonable, it's not perfect but it's a lot better than the busy loop we had before.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2