Update patch set 1
Patch Set 1:

If I understand correctly:

- Either you don't have SMP, and this function is useless (you have only one CPU)
- Or, you have working SMP, and this function is not allowed to fail

So the only case where this is useful is if you have a partially implemented SMP, where you have several CPUs but you don't know how to assign interrupts to them?

If that's the case, I would prefer to not merge this, and instead wait until SMP support on RISC-V is more complete. It is fine as a temporary change for testing things out in the RISC-V port.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>