Update patch set 2
Patch Set 2:

Hi, Since this is a mobile model, INTEL_MODEL_KBYM should be defined and used for this card. This will make the driver/accelerant make the right decisions concerning the internal panel. One thing that will change is the eDP port will be forced (via query IsMobile()) to detect positive on a screen so the driver will actually attempt to program your panel.
At this point the panel will not be detected by EDID(like) code, but it should be actually used.

Patch-set: 2
CC: Gerrit User 1000057 <1000057@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>