Update patch set 1
Patch Set 1:

btw something is not entirely right in the As rule in OverriddenJamRules as it compiles .S files with legacy toolchain. if I put it inside "on $(1)" block (likein Cc and C++) then it picks up the non-legacy toolchain.

BootMergeObject also uses legacy ld which is not quite right.

Patch-set: 1