Update patch set 1
Patch Set 1:

This is very experimental.

The intention: use the same toolchain for .S files as for .c and .cpp.
On x86 currently this is not the case as the bootloader is supposed to be built with non-legacy toolchain. This is true for compiling .cpp files but not for .S files.

Let's see, does it break the CI build.

Patch-set: 1