Update patch set 1
Patch Set 1:

Hello,

Just want to express a word of caution here, hope you don't mind..

While it's of course nice to add support for as many systems as is possible, and I love the energy I see from people for this, I would propose to be a bit carefull since literally -every- new version of intel's cards have differences, making sure just adding ID's will cause trouble for users, when compared to VESA mode, so to speak.
The latest example is for coffeelake and Kabylake, since there I today found out the PLL's are working differently (not only in register locations, but also bits are mangled, pieces are gone and handled differently, and even the calculation itself has to be redone). We are actually very lucky (some) laptops currently seem to work with it on their internal panels.

If we are going to add much ID's I sure hope implementing the belonging functions don't need to be done by me as this will very quickly become more than a day-time job concerning the amount of time it will consume..  ;-)

Patch-set: 1
CC: Gerrit User 1000057 <1000057@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>