Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (2 comments)
> 
> It really looks like a intel_gen8_interrupt_handler would make sense.

Actually (come to think of it) this is why I asked the enum versus bitmask question. I also came to use it the way you propose for ports (portindex), only since the first entry in the index zero is DISPLAY_ANY, I have a +1 sitting in it's use. Also this kind of calculating sits inside the PLL routines.

I was/am a bit worried on top of that, that someone might change to change the portindex to a bitmask too, which would directly break the code.

Of course there will be more than one way to look at it, one could argue this kind of use is a bit of illegal ;-)

Patch-set: 6
Reviewer: Gerrit User 1000057 <1000057@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>