AboutSystem: Add tabs
The goal here is to allow for easier navigation and to
make the information presented in the applet much clearer
and to allow the user to access specific information more
easily while preserving the integrity of their index finger,
if they use a scroll wheel, whether that is a name or figuring
out the dependencies. Some scrolling is okay, too much of it
is laborious. (Even more so, if they are not using a scroll wheel.)

Tabs are the best option, as they are a native add-on that make
navigation much easier, in contrast to other recommendations that
I got, such as drop-down lists and even a table of contents. They
may make the UI "more convoluted", but we're talking about a long
document that is already huge and disorganized. The scrolling window
effect that aims to imitate end credits has not been there in a long
time, so I believe that it is best to reorganize the UI in a pragmatic
manner. The end goal here (as described later) is to achieve a higher
amount of consistency (as far as the UI is concerned), and I believe
that using these other ways are non-standard and contradictory to the
Haiku Interface Guidelines' "Good Software Uses Graphical Controls
Properly" and "Good Software has a Natural Layout to its Controls".
The way that I used tabs here also adheres to the HIG's "Do's and
Don'ts of Tabs".

I anticipated some sort of controversy on this design decision, which
I did kind of expect, as it has been a long time since anyone has touched
it, so I took the liberty to ask random people that were vaguely familiar
with Haiku. I also received feedback from 7 acquaintances about it
by showing both the old version and my version of AboutSystem, and I've
only had one person tell me that "maybe the old one looks more stylish".
I did not clarify which was the change I was working on, and just showed
both versions side-by-side. I explained that the point of the old design
was to imitate the end credits of a movie afterwards, as they did not seem
to understand that, but that did not ultimately affect the answers.

Some other quotes include (No, they are not mine, just trust me on this):
- "I don't think that the credits and the license belong together. I mean
   yes, the authors are related to the copyright, but is that the reason
   why FOSS projects list their contributors? Purely for legal reasons?
   On the other hand, the license is a purely legal matter."
- "I think tabs are nice. And I like credits, in movies, video games and
   operating systems. It would be cool if corporations had a credits
   section for software releases. I'd be like 'Oh, that's my friend!'"
- "Nooo! Don't take away this horribly long scrollbar of unfriendly wall
   text and replace it with organized UI to help you find what you might
   actually care about!" [Safe to say this was sarcastic.]

Most projects separate the licenses and the contributor credits, both in
the GUI and in documents. Some prominent examples are the Xfce Desktop
Environment, FreeBSD, LineageOS, and more. Probably a good idea to imitate
that.

It will also allow for adding more credits that are present in random
BAboutWindow's all over the system without burying the original authors
under a huge wall of text, as well as potentially adding more features to
AboutSystem later down the line, if needed, but that's a whole other debate
that should be dealt with afterwards.

Change-Id: I3e2e9072cd499977dab205d6b358618ba3968a52