Update patch set 1
Patch Set 1:

List of things that I did not do:
a) Add a "system log" tab that showed the system log in real time and included a button to open the /var/log directory as recommended by humdinger, as I also asked PulkoMandy about that and they did not think that this was a good idea. A friend of mine that has some experience in that field also said that "the system log should go to another window" after being shown a screenshot as described in the commit message without me explicitly asking them about that. Maybe later. For now, that seemed to get in the way of getting basically anything done.
b) Move the credits from the separate BAboutWindows (like the ones that I mentioned) to the AboutSystem. I believed that this would add more clutter, for now. This is to be done later on.
c) Move the class definitions to a separate file, as I tried to do that in a separate commit which never got merged. I'll try to do that again, later, in a separate commit.
d) Add the scrolling easter egg back. This would make everything even more confusing, and this already seems like a change that's going to be excessively hard to get through as-is.

This change was shaped around the tons of constructive feedback that I received. I spent more time doing my research on this and authoring a list of arguments than writing the code for it. I am convinced that this is a positive step forward, and even if what we had earlier was fine as well, I think that some improved user experience does not hurt.

Patch-set: 1