Update patch set 1
Patch Set 1: Code-Review-2

Why? If the class is only used in a single .cpp file there is no need to expose it in a .h.

For constants, it will just result in the strings being duplicated in each .o file that includes this .h, which is useless.

So, you are making compile times slower for no reason. There is no rule anywhere which says "class definition must go in .h files" and that is not cleaner, it just exposes internals of a C++ file that shouldn't be exposed.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2