Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> List of things that I did not do:
> a) Add a "system log" tab that showed the system log in real time and included a button to open the /var/log directory as recommended by humdinger, as I also asked PulkoMandy about that and they did not think that this was a good idea. A friend of mine that has some experience in that field also said that "the system log should go to another window" after being shown a screenshot as described in the commit message without me explicitly asking them about that. Maybe later. For now, that seemed to get in the way of getting basically anything done.

I don't think AboutSystem should become more of a "system information" tool than it already is. The current app is already a mix of two things: giving informaitons about Haiku, and about the current machine. I think its main goal is the first of these.

> b) Move the credits from the separate BAboutWindows (like the ones that I mentioned) to the AboutSystem. I believed that this would add more clutter, for now. This is to be done later on.

This would be a worthwhile change and a more obviously useful contribution.

> c) Move the class definitions to a separate file, as I tried to do that in a separate commit which never got merged. I'll try to do that again, later, in a separate commit.

I don't understand why you want to do that. There is no rule that say class definitions must go in separate files, and it is even bad practise since it reduce isolation between different components. Declare things in code to have the smallest visibility possible. It was already discussed in your previous change which was not merged for this reason. You apparently had plans for a "credit roll" screensaver, which would have been a good justification for moving the credits text to a separate header, but the screensaver never materialized.

> d) Add the scrolling easter egg back. This would make everything even more confusing, and this already seems like a change that's going to be excessively hard to get through as-is.

There is a separate change taking care of this, and the discussion there I think was generally agreeing that it wasn't really a great idea. I prefer your idea of a credit roll screensaver.

> 
> This change was shaped around the tons of constructive feedback that I received. I spent more time doing my research on this and authoring a list of arguments than writing the code for it. I am convinced that this is a positive step forward, and even if what we had earlier was fine as well, I think that some improved user experience does not hurt.

I still don't really agree with this.

We spent a long time discussing it privately, I will try to sum up my view on things here.

The AboutSystem text is meant to work similarly to a credit roll in a movie. Not necessarily the rolling, but the idea that it is a long list of names of people and organizations who contributed to Haiku. Because of this, I imagine two main usages for it:
- Showing people that you have your own name in the list, or even, just knowing it's there,
- Reading the whole text from top to bottom (already this is a pretty unlikely scenario)

Your work optimize for a different use case:
- Needing to access a specific part of the credits (the second part with copyright notices for various projects)

I think this use case is very unlikely. Why would people want to access that specific part?

You also consider that the two (or more?) parts of AboutSystem credits are independant, that there are credits on one side, and copyright information in another. To me, this isn't the case. Both parts of these list serve both legal purposes and crediting purposes. On the legal side, the listed names are the copyright holders of all code in Haiku. it doesn't matter if these are individuals or 3rd party projects and organizations. If you wanted to change the license for Haiku to one incompatible with MIT to the point that you need agreement from all contributors, this is a mostly complete list of people you'd need to get agreement from (probably it's missing quite a few). And it is also the place where we can properly credit everyone who contributed in one way or another to Haiku. This is the least we can do, a lot of the developers are unpaid and all they get is the pride of having their name listed there. This also applies to both individual people and organizations/3rd party projects. So I really understand this as one single document.

Following that thinking, the idea of splitting the document into two parts makes little sense to me. Moreover, I think tabs are not a great user interface, because they are small and hard to reach and limited in what they can do. In fact they are so limited that you will still need a scrollbar in each of the tabs. I think the document is easier to understand as a single long document (like a webpage, or a pdf file, or any thing of that kind). This is a metaphor people understand well, and the idea of scrolling is so popular that the mouse was augmented with a dedicated scroll wheel to make it very easily reachable. Or, for more precise control you can also use the scrollbar in several ways (it is a quite advanced control, with click, hold, and drag of the mouse on different areas having different effects).


To sum it up, the "pros" would be:
- It covers a new use case
- It makes the structure of the data more explicit in the user interface

The combination of these three things make me quite unconvinced that this is a good change:
- It makes the UI more complicated (adding tabs and not removing anything)
- It splits into two a document that to me is a single all-in-one thing (in other words, the structure it makes visible is not a relevant one)
- The use case it tries to handle (accessing a specific subsection of the document) seems unlikely to ever be needed by anyone


And new information that we didn't discuss yet:

In the commit message you mention that you have followed the user interface guideliens about tabs. Since you bring that in, I checked it and you went against at least one of the rules: "Don't use a scrollbar to scroll a form inside a tab" (ok, this isn't really a form, but I think the general idea still applies).


That being said, I have now spent hours discussing this and I don't think it is worth spending more of my time for the change on its own. And I'm not here to pick a fight over a part of Haiku I hardly ever look at anyway. If someone thinks this is cool and merges it, I'm fine with that, it doesn't really matter to me. The reason I spend time discussing this is so we can all learn and think more about UI design and use our better understanding when working on future (hopefully more meaningful/relevant?) changes to the UI. Also I suspect there is some bikeshed effect: the more you decide to work on the small and seemingly trivial things, the more people will have an opinion on your work and want to take part in the discussion (https://personalexcellence.co/blog/bike-shed-effect/). Probably combined with a touch of resistance to change?

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>