Update patch set 2
Patch Set 2: Code-Review-1

I think it is more than safe to say this is already a bikeshed based on how long the comments on this change are.

I am in agreement with PulkoMandy that tabs are not an improvement here and the existing setup is fine. Here are some thoughts from me about that:

 * If using the scroll wheel is too much work (it may well be here), that's what the scroll bar and the Page Up/Down keys are for. Ahh, much better.

 * This screen is probably not going to be looked at for more than about 30 seconds for most users, and they aren't going to bother scrolling and reading everything. Even if we put it into tabs, they probably still won't bother looking at it, at all. This information is just not useful to 99% of people who use Haiku, and so they won't bother with it. Spending hours optimizing the information is just not a good use of time in my book. Further, tabs make it difficult to Cmd+A, Cmd+C, and then paste into a text editor/browser of your choice.

 * If we want to reinstate the "credits scroll", that's fine with me, I think, and would be preferable to tabs.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1