Update patch set 2
Patch Set 2:

> I think it is a very normal thing for me to be wary of taking on progressively greater projects and to progressively ask little, annoying questions all the time until I get burned out. I think I can't get more honest when I say that I have zero sense of direction here, and when I try to compromise my changes according to the directions of other people, it completely backfires, like the part where my change was perceived as removing the significance of third-party dependencies as a result of me adjusting the wording.

I hear your fears here, and I think that means we (i.e. you and us) need to find a better way of communicating about expectations and what changes we might find valuable, and why. Bigger endeavors always warrant discussion on the mailing lists, smaller ones just on Trac, and very small ones maybe then just on IRC or Discourse; but things like this which are UX-related are bound to generate large bikesheds (well, this one already has) and just asking a few people for their thoughts may not be "sufficient" to determine that the change would be generally accepted.

I also think "it's fine as is" is a genuinely valid response here. I see you have laid out various ways your new approach is an improvement, but from my point of view, those things are not actually things we need or even want to optimize for here. But I think PulkoMandy's outline explains that "it's fine as is" in such detail, so, I don't we are being "dismissive" or "trite" here.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>