Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > So I see 3 options
> > > A. change riscv64 to use the newly introduced interrupt_controller field
> > > B. change arm to use its own field (e.g. arch_args.gic) and then we can remove interrupt_controller field
> > > C. change both to do probing in the kernel and then no such field is needed in arch_args. This is possible if we can do fdt parsing in arch_int_init() or arch_int_init_post_vm() - or perhaps arch_int_init_post_device_manager() ?
> > 
> > B.2 move interrupt_controller to `arch/arm` and only handle it in ARM code. As I understand multiple controller types need to be handled in ARM, not only GIC.
> > 
> > The problem with A is that RISC-V will need its own private interrupt controller fields like total interrupt count. So per-arch interrupt_controller struct version is needed.
> > 
> > The problem with C is that FDT bus is not yet avalible at moment of initalizing interrupts. Making 2 FDT bus drivers in kernel is probably bad idea.
> 
> Yes, it makes sense to have some variant of B then (move the ARM interrupt controller into ARM specific files). All architectures have an interrupt controller, but the inner details of it are probably different.

That's the rub.  This interrupt_controller field is already in architecture specific headers. The solution would be to create a dtb parsing file for every architecture (this field today is populated by dtb.cpp in efi.  That would result in a lot of duplicate code.

> This patch is not great because it adds an unused field to risc-v and that adds some confusion.

Technically it can be used however... kernel args are passed to the init calls for the interrupt controller. There's even a todo there to look for interrupt controller info in fdt (which this patch does... just in the bootloader passing to kernel args)

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>