Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> https://review.haiku-os.org/c/haiku/+/4774

It's quite similar to what I started with in the first version of my patch for interrupt controller detection. But the drawback there is excessive use of "ifdef arm" which as I understand is discouraged in Haiku. Still, for me that's a viable alternative at least in the short term.

Patch-set: 1
Reviewer: Gerrit User 1000341 <1000341@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>