Update patch set 1
Patch Set 1:

> The "real" solution is to redesign all of our interrupt controller code to be an add-on / bus based instead of something the kernel does statically.  However the riscv64 build would be broken until that happens which is undesirable.

This would be nice, but is not realistic. Interrupts need to be set up early on in the kernel because basically everything else needs them working. So it needs to be set up statically, with the bootloader feeding the needed info in an easy to use way.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>