Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > The "real" solution is to redesign all of our interrupt controller code to be an add-on / bus based instead of something the kernel does statically.  However the riscv64 build would be broken until that happens which is undesirable.
> > 
> > This would be nice, but is not realistic. Interrupts need to be set up early on in the kernel because basically everything else needs them working. So it needs to be set up statically, with the bootloader feeding the needed info in an easy to use way.
> 
> So would the alternative be moving the fdt parsing out of the bootloader and into the kernel?  I feel like we did this originally with the fdt helpers, but they were removed when the fdt bus stuff was merged.

I don't see how that will help.

You need to do some early parsing of the FDT at some point to get the UART and interrupt controller. At least the UART is probably useful to the bootloader itself in some cases, and both are useful to the kernel *very early* (really it is the first things the kernel has to do).

So, this is what we do:
- The bootloader does some limited FDT parsing to locate the uart and interrupt controller, and feeds this info to the kernel, in platform specific ways because this is platform specific
- The kernel use this info for the early boot
- Later on, the kernel does its own FDT parsing (once it is set up, with memory allocation, busses, drivers, etc) to find all other devices.

This architecture is good, there is no need to change it.

The disagreement between ARM and RISC-V is on where to put the info about the interrupt controller. RISC-V put it in a platform specific file. ARM currently puts it in a platform-shared file, but guards it with a #if. I wasn't sure if we were going to proceed with making the code shared between platforms as proposed in the ARM implementation. Apparently, that is not possible or desirable, because the data needed is too specific to each platform. So now the ARM port should be adjusted to work similarly to RISC-V, with these fields passed to the kernel in architecture-specific data (and out of the sourcecode that is shared with other platforms).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>